requirements.txt
absl-py==2.3.0
ale-py==0.11.1
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.2
cryptography==43.0.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
Farama-Notifications==0.0.4
fonttools==4.58.1
google-auth==2.40.2
google-genai==1.18.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
ipython==8.37.0
ipywidgets==8.1.7
jedi==0.19.2
jiter==0.10.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
ml_collections==1.1.0
multi-agent-ale-py==0.1.11
numpy==1.26.4
open_spiel==1.6
openai==1.83.0
opencv-python==4.11.0.86
overcooked_ai==1.1.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pettingzoo==1.25.0
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
stack-data==0.6.3
SuperSuit==3.10.0
tinyscaler==1.2.8
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
volcengine-python-sdk==3.0.7
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.14
